+++ packages/core/src/errors.ts
@@ -1,7 +1,0 @@
-import { Data } from "effect"
-
-export class QuartzError extends Data.TaggedError("QuartzError")<{
-  readonly message: string
-  readonly cause?: unknown
-}> {}
-